requirements.txt
pyspark==3.5.0
requests==2.31.0
pandas==2.1.4
kafka-python==2.0.2
influxdb-client==1.36.1
python-dotenv==1.0.0 